armv8 build scripts update
